+++ src/Monify/Semantics/INamedTypeSymbolExtensions.ToSubject.cs
@@ -27,10 +27,10 @@ internal static partial class INamedTypeSymbolExtensions
     /// <param name="value">
     /// The type of the value encapsulated by the <see cref="Subject"/>.
     /// </param>
-    /// <param name="generatePassthrough">
+    /// <param name="passthrough">
     /// A value indicating whether passthrough interfaces, methods and properties should be generated.
     /// </param>
-    /// <param name="generateDebuggerDisplay">
+    /// <param name="debuggerDisplay">
     /// A value indicating whether the debugger display should be generated.
     /// </param>
     /// <returns>
@@ -45,8 +45,8 @@ internal static partial class INamedTypeSymbolExtensions
         SemanticModel model,
         ImmutableArray<Nesting> nesting,
         ITypeSymbol value,
-        bool generatePassthrough = true,
-        bool generateDebuggerDisplay = true)
+        bool debuggerDisplay = true,
+        bool passthrough = true)
     {
         string @namespace = subject.ContainingNamespace.IsGlobalNamespace
            ? string.Empty
@@ -67,8 +67,8 @@ internal static partial class INamedTypeSymbolExtensions
             CanOverrideGetHashCode = subject.CanOverrideGetHashCode(),
             CanOverrideToString = subject.CanOverrideToString(),
             Declaration = declaration,
-            Encapsulated = subject.GetEncapsulated(compilation, model, value, generatePassthrough),
-            GenerateDebuggerDisplay = generateDebuggerDisplay,
+            Encapsulated = subject.GetEncapsulated(compilation, model, value, passthrough),
+            GenerateDebuggerDisplay = debuggerDisplay && !subject.HasDebuggerDisplay(),
             HasEqualityOperator = subject.HasEqualityOperator(),
             HasEquatable = subject.HasEquatable(),
             HasField = hasFieldForEncapsulatedValue,